+++ src/ProjectConfig/ProjectConfig.php
@@ -1507,8 +1507,12 @@ class ProjectConfig
 
             if (! empty($projectConfigNames)) {
                 foreach ($projectConfigNames as $uid => $name) {
-                    $uids[] = '/^(.*'.preg_quote((string) $uid).'.*)$/mi';
-                    $replacements[] = '$1 # '.$name;
+                    $name = trim((string) $name);
+
+                    if ($name !== '') {
+                        $uids[] = sprintf('/^.*\b%s\b.*$/m', preg_quote((string) $uid));
+                        $replacements[] = "$0 # $name";
+                    }
                 }
             }