+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Users can now connect their accounts to one or more Socialite providers. ([#19202](https://github.com/craftcms/cms/pull/19202))
+- The login page now lists Socialite providers. ([#19202](https://github.com/craftcms/cms/pull/19202))
+- Fixed a bug where some control panel resources and pages weren’t loading properly. ([#19214](https://github.com/craftcms/cms/issues/19214))
+
+## 6.0.0-alpha.10 - 2026-07-03
+
 - It’s now possible to load Blade views within Twig templates. ([#19148](https://github.com/craftcms/cms/pull/19148))
 - Craft now registers several Blade directives, bringing near feature parity with Twig templating. ([#19148](https://github.com/craftcms/cms/pull/19148))
 - Added SQLite database support. ([#19149](https://github.com/craftcms/cms/pull/19149))
@@ -12,6 +18,8 @@
 - Fixed a bug where the login page would show a CSRF token mismatch on http requests
 - Fixed an error that could occur during Craft 6 upgrades when the `migrations` table was missing its `track` column. ([#19168](https://github.com/craftcms/cms/pull/19168))
 - Fixed an error that occurred when saving Users source settings after selecting the Groups table column. ([#19184](https://github.com/craftcms/cms/pull/19184))
+- Fixed a bug where database backups weren’t using the `--single-transaction` or `--column-statistics=0` flags on MySQL.
+- Fixed a bug where Content Block fields’ nested fields weren’t available as table columns on element indexes.
 
 ## 6.0.0-alpha.9 - 2026-06-23